Fix address err

+++ multi_vendor_user/lib/Pages/product_detail.dart
@@ -436,9 +436,7 @@ class _ProductDetailsPageState extends State<ProductDetailsPage> {
                       child: Row(
                         children: [
                           RatingBarIndicator(
-                            rating: totalUser == 0
-                                ? 0
-                                : getRatingAndReview().toDouble(),
+                            rating: totalUser == 0 ? 0 : getRatingAndReview(),
                             itemBuilder: (context, index) => const Icon(
                               Icons.star,
                               color: Colors.orange,
